add team views and serializers

+++ user/views/profile.py
@@ -1,15 +1,14 @@
-from rest_framework.permissions import IsAuthenticated
 from rest_framework.views import APIView
+from rest_framework.permissions import IsAuthenticated
 
-from rest_framework_simplejwt.authentication import JWTAuthentication
 from rest_framework.response import Response
 from user.models import User
 from user.authentication import CookieTokenAuthentication
-from user.serializers import UserSerializer
+from user.serializers import UserSerializer, DeleteProfileSerializer
 
 
 class ProfileView(APIView):
-    authentication_classes = [CookieTokenAuthentication, JWTAuthentication]
+    authentication_classes = [CookieTokenAuthentication]
     permission_classes = [IsAuthenticated]
 
     def get(self, request, *args, **kwargs):
@@ -19,7 +18,7 @@ class ProfileView(APIView):
 
 
 class UpdateProfileView(APIView):
-    authentication_classes = [CookieTokenAuthentication, JWTAuthentication]
+    authentication_classes = [CookieTokenAuthentication]
     permission_classes = [IsAuthenticated]
     serializer_class = UserSerializer
 
@@ -56,10 +55,12 @@ class UpdateProfileView(APIView):
 
 
 class DeleteAccountView(APIView):
-    authentication_classes = [CookieTokenAuthentication, JWTAuthentication]
+    authentication_classes = [CookieTokenAuthentication]
     permission_classes = [IsAuthenticated]
+    serializer_class = DeleteProfileSerializer
 
     def post(self, request, *args, **kwargs):
+        # TODO: add more logic incase if the team is already registered for a hackathon
         user = request.user
         password = request.data.get('password')
         if not user.check_password(password):
@@ -69,6 +70,20 @@ class DeleteAccountView(APIView):
                     'message': 'Invalid password'
                 }
             }, status=400)
+        from team.models import TeamMember
+        if TeamMember.objects.filter(user=user, isOwner=True).exists():
+            return Response({
+                'error': {
+                    'code': 'TEAM_OWNER',
+                    'message': 'You are the owner of a team. Please delete the team or transfer ownership to someone else'
+                }
+            }, status=400)
+        if TeamMember.objects.filter(user=user).exists():
+            team_member = TeamMember.objects.filter(user=user).first()
+            team = team_member.team
+            team_member.delete()
+            if team.teammember_set.count() == 0:
+                team.delete()
         user.delete()
         return Response(status=200)
